// superset-frontend/plugins/plugin-chart-elevator-status/src/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
import { QueryFormData, TimeseriesDataRecord } from '@superset-ui/core';

export interface ElevatorStatusStylesProps {
  height: number;
  width: number;
}

export type DirectionType = 'up' | 'down' | 'idle';

export interface ElevatorStatusCustomizeProps {
  // Column Mappings
  statusColumn?: string;
  floorColumn?: string;
  directionColumn?: string;
  payloadColumn?: string;
  destinationColumn?: string;

  // Customizations
  doorColor: string;
  frameColor: string;
  arrowColor: string;
  showDirectionArrows: boolean;
  showFloorDisplay: boolean;
  statusText?: string;
  fontSize: number;
}

export interface ElevatorStatusProps extends ElevatorStatusStylesProps {
  data: TimeseriesDataRecord[];
  customize: ElevatorStatusCustomizeProps;
}

export type ElevatorStatusQueryFormData = QueryFormData &
  ElevatorStatusStylesProps &
  ElevatorStatusCustomizeProps;